==> notebooks/3_Demonstrativo_Resultado.ipynb <==
import pandas as pd
import os

# 1. Ler Arquivos

# Sobe um nível no diretório
os.chdir("..")

# Armazena o novo diretório de trabalho atual
diretorio_atual = os.getcwd()
print(diretorio_atual)

# Caminho dos diretorios
dir_preprocessed_data = os.path.join(diretorio_atual, "data_cvm\preprocessed_II")

# Define o diretório
os.chdir(dir_preprocessed_data)

# Ler dados trimestral
df_trim = pd.read_csv("equatorial_resultado_trim.csv")

# Ler dados anual
df_anual = pd.read_csv("equatorial_resultado_anual.csv")

df = pd.concat([df_anual, df_trim], ignore_index=True)

Os dados trimestrais e anuais foram concatenados para incluir o 4 trimestre, que será calculado posteriormente. \ 
O processamento deve ser realizado em ambos os dados, anuais e trimestrais(concatenado).

# 2. Análise Exploratória

## 2.1 Verficando se há meses duplicados 

teste = df[(df['CD_CONTA'] == '3.01') & (df['ANO'] == 2022)] 

display(teste)

Os meses 6 e 9 possuem valores onde o exercício começa em janeiro do ano vigente e valores onde começam no fim do trimestre anterior.

# 3. Manipulação dos Dados

## 3.1 Filtrando as Contas Por Nível

# Filtra contas dos níveis mais altos
df_anual = df_anual[df_anual['CD_CONTA'].str.len() <= 7]
df = df[df['CD_CONTA'].str.len() <= 7]

## 3.1 Deletando os meses 6 e 9 com inicio do exercicio em janeiro

### 3.1.1 Criando coluna ini_mes

# Converte coluna DT_REFER para o tipo datetime
df['DT_INI_EXERC'] = pd.to_datetime(df['DT_INI_EXERC'])

# Separa o mês e cria a coluna mês
df['INI_MES'] = pd.DatetimeIndex(df['DT_INI_EXERC']).month

# Crie uma condição para identificar as linhas que deseja excluir
condicao = (df["MES"].isin([6, 9])) & (df["INI_MES"] == 1)
# print(condicao)

df_droped = df.drop(df[condicao].index)
# display(df_drop.tail())

# VALIDAÇÃO DA EXCLUSÃO - resultado vazio esta ok
print(df_droped[(df_droped['MES'] == 9) & (df_droped["INI_MES"]== 1)])

teste = df_droped[(df_droped['CD_CONTA'] == '3.01') & (df_droped['ANO'] == 2022)] 
display(teste)

## 3.2 Deletando as Colunas

df_droped_column_anual = df_anual.drop(
    columns=['DT_FIM_EXERC','DT_INI_EXERC'])

df_droped_column = df_droped.drop(
    columns=['DT_FIM_EXERC', 'INI_MES','DT_INI_EXERC'])

## 3.3 Inserindo a Coluna Período

### 3.3.1 Anual

df_droped_column_anual['PERIODO'] = 'ANUAL'

### 3.3.2 Trimestral

# Função para inserir a coluns Período
def insere_periodo_trimestral(df):
    df = df.copy()

    # Dicionário para mapear os meses aos trimestres
    mes_para_trimestre = {3: 1, 6: 2, 9: 3, 12: 4}
    
    # Filtrar apenas os anos e meses relevantes
    anos_relevantes = list(range(2011, 2025))
    meses_relevantes = [3, 6, 9, 12]

    # Aplicar a lógica de período trimestral
    df = df[(df['ANO'].isin(anos_relevantes)) & (df['MES'].isin(meses_relevantes))]
    df['PERIODO'] = df.apply(lambda row: f"{mes_para_trimestre[row['MES']]}T{row['ANO']}", axis=1)
    
    return df

# Inserindo a coluna período
df_droped_column = insere_periodo_trimestral(df_droped_column)

df_droped_column[(df_droped_column["ANO"] == 2023) & (df_droped_column['CD_CONTA'] == "3.02")]

## 3.4 Calculando 4º Trimestre

# Cria uma cópia do dataframe
df_calculado = df_droped_column.copy()

# Obter os anos e contas únicos
anos = df_calculado['ANO'].unique()
contas = df_calculado['CD_CONTA'].unique()

# Soma os valores do mes 3, 6 e 9 e subtrai o valor do mês 12 
for ano in anos:
    for conta in contas:
        # Filtra o DataFrame com base nas condições de conta e ano
        filtro = (df_calculado['CD_CONTA'] == conta) & (df_calculado['ANO'] == ano)
        
        # Filtra as linhas onde o MES é igual a 3, 6 ou 9
        filtro_mes_369 = filtro & df_calculado['MES'].isin([3, 6, 9])
        
        # Obtém a soma dos valores da coluna VL_CONTA onde MES é igual a 3, 6 ou 9
        soma_mes_369 = df_calculado[filtro_mes_369]['VL_CONTA'].sum()
        
        # Filtra as linhas onde o MES é igual a 12
        filtro_mes_12 = filtro & (df_calculado['MES'] == 12)
        
        # Subtrai a soma dos meses 3, 6 e 9 do valor do mês 12
        df_calculado.loc[filtro_mes_12, 'VL_CONTA'] = df_calculado.loc[filtro_mes_12, 'VL_CONTA'] - soma_mes_369

print("DataFrame atualizado:")
# display(df_drop_column.head())

df_calculado[(df_calculado["ANO"] == 2022) & (df_calculado['CD_CONTA'] == "3.02")]

Resultado esperado: \
1T2022 = -3844293.0 \
2T2022 = -4380488.0 \
3T2022 = -4882854.0 \
4T2022 = -5458101.0	

## 3.5 Concatenando Dados Anuais e Trimestrais

df_final = pd.concat([df_calculado, df_droped_column_anual], ignore_index=True)

## 3.6 Alterando o Nome das Colunas

colunas = ['CNPJ','EMPRESA','CONTA','DESCRIÇÃO','VALOR','DEMONSTRATIVO','ANO', 'MES', 'PERIODO']
df_final.columns = colunas

# 4. Exportando os Dados

# Define o diretório de dados processados
dir_processed_data = os.path.join(diretorio_atual, "data_cvm\processed")
os.chdir(dir_processed_data)

df_final.to_csv('demonstracao_resultado_completo.csv', index=False)
